Remove ASTLinter, rename BaseASTLinter => ASTLinter
This was a simplifiation step for the common case; that simplification
has now been applied to all cases.
