+++ tests/fixtures/MustUseOverrideAttributeLinter/with_generic_parent.php.expect
@@ -2,6 +2,6 @@
     {
         "blame": "  public function getPrettyTextForNode(EditableNode $node): EditableNode {\n  }\n",
         "blame_pretty": "  public function getPrettyTextForNode(EditableNode $node): EditableNode {",
-        "description": "Foo::getPrettyTextForNode() overrides Facebook\\HHAST\\Linters\\BaseASTLinter::getPrettyTextForNode() without <<__Override>>"
+        "description": "Foo::getPrettyTextForNode() overrides Facebook\\HHAST\\Linters\\ASTLinter::getPrettyTextForNode() without <<__Override>>"
     }
 ]
